Generalize test harness to accommodate stdout-dependent extra file names
The forthcoming Orca parsers require stdout in order to determine the
names of the extra files. The previous test harness set-up was
incompatible with that, because:

1. `_test_parser_direct()` assumed `contents` was the file to be parsed,
   which for the Hessian required me to set, for example,
   `contents=Path("water.hess.hess")`

2. This caused `_test_decode_integration()` to fail because it didn't
   have stdout, which is needed to locate the correct `.hess` file

The new test harness implementation avoids this by simply passing
`stdout` to the test case harness (can be `None` if not needed), and
determining which of the extra files is the parser content from the
parser registry, similar to what decode does internally.